codebuddy/ai_agents/pyproject.toml
[project]
name = "ep-ai-agents"
version = "0.1.6"
description = "ai agents"
readme = "README.md"
classifiers = ["Private :: Do Not Upload"]
requires-python = ">=3.12.8,<3.13"
dependencies = [
    "openinference-instrumentation-smolagents>=0.1.4",
    "smolagents[litellm,mcp,openai,telemetry]==1.21.1",
    "gitpython>=3.1.44",
    "aiohttp>=3.11.11",
    "chardet>=5.0.0",
    "tree-sitter>=0.24.0,<0.25.0",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-c==0.23.1",
    "tree-sitter-cpp>=0.23.4",
    "litellm>=1.72.0",
    "openai<1.100.0",
    "opentelemetry-sdk>=1.31.1",
    "opentelemetry-exporter-otlp>=1.31.1",
    "langfuse>=2.0.0,<3.0.0",
    "fire>=0.7.0",
    "pyyaml>=6.0.0",
    "bashlex>=0.18",
    "tenacity>=9.1.2", #这是个retry的包，litellm 的 retry 逻辑依赖这个
    "rich>=13.9.4",
    "arrow>=1.3.0",
    "datasets>=3.6.0",
    "packaging>=24.2",
    "sentry-sdk>=2.35.1",
    "ruamel-yaml>=0.18.10", # packages/commonlibs/commonlibs/encoding/xyaml.py内依赖
]

[project.scripts]
ept_simple_cli = "clis.simple.cli:main"
ept_haloos_cli = "clis.haloos_unit_test.cli:main"

[project.optional-dependencies]
haloos = [
    "beautifulsoup4>=4.12.0",
    "gcovr>=8.3",
    "libclang",
]
simple = []
swe = [
    "pandas",
]

[[tool.uv.index]]
url = "https://artifactory.ep.chehejia.com/artifactory/api/pypi/pypi-remote/simple"
default = true

[[tool.uv.index]]
name = "liauto"
url = "https://artifactory.ep.chehejia.com/artifactory/api/pypi/liauto-pypi-l5/simple"
publish-url = "https://artifactory.ep.chehejia.com/artifactory/api/pypi/liauto-pypi-l5"
explicit = true

[tool.uv.sources]
externalapis = { path = "../../packages/externalapis", editable = true }
sysutils = { path = "../../packages/sysutils", editable = true }
commonlibs = { path = "../../packages/commonlibs", editable = true }
repoutils = { path = "../../packages/repoutils", editable = true }

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
testpaths = ["tests"]

[tool.ruff.lint.isort]
known-first-party = ["ai_agents"]
known-third-party = ["commonlibs", "externalapis", "repoutils", "sysutils"]


[dependency-groups]
dev = [
    "commonlibs",
    "externalapis",
    "repoutils",
    "sysutils",
    "gradio>=5.23.3",
    # 因为这个问题，目前 fire 要求 ipython < 9.0 https://github.com/google/python-fire/pull/588
    "ipython<9.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.26.0",
    "responses>=0.25.0",  # For HTTP request mocking in tests
]
eval = [
    "datasets>=3.6.0",
    "docker>=7.1.0",
    "swebench>=4.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["ai_agents", "clis", "sysutils", "externalapis", "commonlibs", "repoutils"]
